refactor: reorganize modules and split parsing from manipulation
